Bearing calculation reverse last point direction
In the rating implementation, when do the last point bearing calculation, do we need to reverse the direction?
Recently, we found the matching rate has some trouble, and after debug into, found the issue is caused by this direction change for last point. 
If we remove this line, the MAX_BEAR_DIFF can set as normal, otherwise, we have to set as 180. 
So I believe we should remove this line, but not sure what is the reason we put it in there